re-organize components into folders, also user data returned from backend need to be specify in front-end mutations for login and add user and in queries when using them
