+++ client/src/App.tsx
@@ -10,9 +10,9 @@ import { Outlet } from 'react-router-dom';
 import { ThemeProvider } from 'styled-components';
 import theme from './theme';
 import GlobalStyles from './GlobalStyles';
-import Container from './components/Container';
-import Background from './components/Background';
-import AddCapsuleService from './components/AddCapsuleService';
+import Container from './components/layout/Container';
+import Background from './components/layout/Background';
+import AddCapsuleService from './components/api/AddCapsuleService';
 
 // Construct our main GraphQL API endpoint
 const httpLink = createHttpLink({
