+++ client/src/components/common/FloatingButton.tsx
