+++ client/src/components/layout/Background.tsx
