+++ client/src/components/layout/CapsuleLayout.tsx
