+++ client/src/components/layout/Footer.tsx
