+++ client/src/components/layout/Header.tsx
@@ -1,8 +1,8 @@
 import { Link } from 'react-router-dom';
 import { type MouseEvent} from 'react';
-import Auth from '../utils/auth';
-import Navbar from './Navbar';
-import StyledButton from './StyledButton';
+import Auth from '../../utils/auth';
+import NavbarLayout from './NavbarLayout';
+import StyledButton from '../common/StyledButton';
 
 const Header = () => {
   const logout = (event: MouseEvent<HTMLButtonElement>) => {
@@ -16,7 +16,7 @@ const Header = () => {
   };
   return (
     <header>
-      <Navbar>
+      <NavbarLayout>
         <div>
           <Link style={{color: '#C466FF'}} to="/">
             <h1>Capsule Time</h1>
@@ -37,7 +37,7 @@ const Header = () => {
             </li>
           </ul>
         </div>
-        </Navbar>
+        </NavbarLayout>
     </header>
   );
 };
