+++ client/src/pages/Profile.tsx
@@ -1,5 +1,5 @@
-import Footer from '../components/Footer';
-import Header from '../components/Header';
+import Footer from '../components/layout/Footer';
+import Header from '../components/layout/Header';
 
 //This will be capsules
 // import ThoughtForm from '../components/ThoughtForm';
@@ -21,7 +21,7 @@ const Profile = () => {
           <div>
           <h2>Profile Page</h2>
           <p>manage capsule - delete, not allowed to open (blurry)</p>
-          <div>display unlock capsules</div>
+          <div>display user capsules</div>
           </div>
         ) : ( 
           <p>you need to be logged in to see this content!</p>
